postprocess: Move /etc/ld.so.cache → /usr
This is an "index file" similar to many others; see for example
[the fontconfig case](https://bugzilla.redhat.com/show_bug.cgi?id=1377367).
That one was in `/var`, and while `/etc` isn't quite as bad,
it's still not a human-editable file, and should be lifecycle bound
with `/usr`.

This will be a bit more important with livefs, since with that we'd need to
change the running `/etc`.