tap-test: fix typo and actually create .test marker
We have this chunk in tap-test:

    if test -f ${test_tmpdir}/.test; then
      rm "${tempdir}" -rf
    fi

But (1) we were never actually touching `.test` and (2), we were
checking in `test_tmpdir` instead of `tmpdir`. This meant that the test
dirs were never actually cleaned up.

Create the marker file and make sure we rm the right dir. Also re-indent
to remove tabs.