bin/compose: Suppress kernel.rpm's dracut run
Finally got around to implementing the bit from
https://github.com/systemd/systemd/pull/4174
however suboptimal it is; need the unified core
so we can cleanly ignore the posttrans like
we do others.

Note I intentionally keep the file around in the generated tree so that
installing a kernel RPM per client doesn't try to do any of this either.