compose: Clean up internal final postprocess API
It's really way saner if we create the target root dir in the
compose high level.  That way everything relating to actual file names
(and the concept of the "workdir") rather than the "library".

This is prep for [compose split](https://github.com/projectatomic/rpm-ostree/pull/1039)
in that it will be easier to tease out an API for less "opinionated"
postprocessing that just works on `target_rootfs_dfd`.